Import Features component directly from lib
